Beginning work on the size == 4 solver. Looks like this one may be brute-forceable.

+++ tomstuff/src/Game/GameClient/ViewModels/Problem.cs
@@ -9,6 +9,6 @@ namespace GameClient.ViewModels
         public long Size { get; set; }
         public List<string> Operators { get; set; }
         public bool Solved { get; set; }
-        public long TimeLeft { get; set; }
+        public double TimeLeft { get; set; }
     }
 }
